Fix readme generation in build-homepage
Quoting the `find` in for can make bash not split on newline and give
an error like below. This should fix that. Also switched to a while
loop with -print0 so as to handle file names with spaces.

```
Generating the README.md
Adding the member: ./crew/jane-doe.md
./crew/meain.md
cat: './crew/jane-doe.md'$'\n''./crew/meain.md': No such file or directory
All crew members are updated!
```

Signed-off-by: Abin Simon <mail@meain.io>